rename: 마이페이지 관련 컴포넌트 파일 이동
